+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 0.4.5 (unreleased)
 ==================
 
+* Incompatible Change: Renamed `padl` and `padr` to `pad_start` and `pad_end`
+to make it more consistent naming to `trim_start` and `trim_end`.
 * Change: Added proper error location to errors generated from
 prelude or other pure Rust functions.
 * Change: Added recently defined variable name or key to syntax positions